refactor: Separate design-time stubs into individual files and fix implementations
- Moved DesignTimeDbContextFactoryInstance, DesignTimeDatabaseProvider,
  and DesignTimeLockingBehavior to their own files to comply with SA1402.
- Corrected nullability for IJellyfinDatabaseProvider.DbContextFactory
  in the DesignTimeDatabaseProvider stub to resolve CS8767 and CS8618.
- Ensured all interface members are stubbed to prevent CS0535 errors.
- Addressed various StyleCop formatting issues within these files.
